+++ backend/services/auth-service/main.go
@@ -1,13 +1,13 @@
 package main
 
 import (
-	"github.com/rolanmulukin/tatar-shower-backend/db"
 	"log"
 	"net/http"
 	"os"
 	"time"
 
-	"github.com/gorilla/mux"
+	"github.com/rolanmulukin/tatar-shower-backend/db"
+
 	"github.com/rolanmulukin/tatar-shower-backend/services/auth-service/handlers"
 )
 
@@ -22,11 +22,7 @@ func main() {
 	}
 
 	h := handlers.NewHandler(sqlDB)
-	r := mux.NewRouter()
-	r.Use(corsMiddleware)
-	api := r.PathPrefix("/api").Subrouter()
-	api.HandleFunc("/register", h.RegisterUser).Methods("POST")
-	api.HandleFunc("/signin", h.SingInUser).Methods("POST")
+	r := h.SetupRoutes()
 
 	srv := &http.Server{
 		Addr:         ":8001",
@@ -36,19 +32,4 @@ func main() {
 	}
 	log.Println("auth-service starting on", srv.Addr)
 	log.Fatal(srv.ListenAndServe())
-}
-
-func corsMiddleware(next http.Handler) http.Handler {
-	return http.HandlerFunc(func(w http.ResponseWriter, r *http.Request) {
-		w.Header().Set("Access-Control-Allow-Origin", "*")
-		w.Header().Set("Access-Control-Allow-Methods", "GET, POST, PUT, DELETE, OPTIONS")
-		w.Header().Set("Access-Control-Allow-Headers", "Content-Type, Authorization")
-
-		if r.Method == http.MethodOptions {
-			w.WriteHeader(http.StatusNoContent)
-			return
-		}
-
-		next.ServeHTTP(w, r)
-	})
 }
