+++ backend/services/rec-service/main.go
@@ -1,22 +1,19 @@
 package main
 
 import (
-	"github.com/gorilla/mux"
-	"github.com/rolanmulukin/tatar-shower-backend/models"
-	"github.com/rolanmulukin/tatar-shower-backend/services/rec-service/handlers"
 	"log"
 	"net/http"
 	"time"
+
+	"github.com/rolanmulukin/tatar-shower-backend/models"
+	"github.com/rolanmulukin/tatar-shower-backend/services/rec-service/handlers"
 )
 
 func main() {
 	// TODO: switch to DB-backed storage
 	storage := models.NewStrorage()
 	h := handlers.NewHandler(storage)
-	r := mux.NewRouter()
-	api := r.PathPrefix("/api").Subrouter()
-	api.HandleFunc("/stats/streak", h.GetStreakHandler).Methods("GET")
-	api.HandleFunc("/tips/random", h.GetTipsHandler).Methods("GET")
+	r := h.SetupRoutes()
 
 	srv := &http.Server{
 		Addr:         ":8003",
