+++ backend/services/shower-service/main.go
@@ -1,7 +1,6 @@
 package main
 
 import (
-	"github.com/gorilla/mux"
 	"github.com/rolanmulukin/tatar-shower-backend/models"
 	"github.com/rolanmulukin/tatar-shower-backend/services/shower-service/handlers"
 	"log"
@@ -13,12 +12,7 @@ func main() {
 	// TODO: switch to DB-backed storage
 	storage := models.NewStrorage()
 	h := handlers.NewHandler(storage)
-	r := mux.NewRouter()
-	api := r.PathPrefix("/api").Subrouter()
-	api.HandleFunc("/user/schedules", h.GetAllSchedulesHandler).Methods("GET")
-	api.HandleFunc("/user/schedules", h.CreateOrUpdateScheduleHandler).Methods("POST", "PUT")
-	api.HandleFunc("/user/schedules", h.DeleteScheduleHandler).Methods("DELETE")
-	api.HandleFunc("/user/shower/completed", h.CompleteShowerHandler).Methods("POST")
+	r := h.SetupRoutes()
 
 	srv := &http.Server{
 		Addr:         ":8002",
